transfer-learning/indoor-outdoor/requirements.txt
streamlit
tensorflow
Pillow
requests
huggingface-hub